Allow duplicating collection from detail page
